requirements.txt
# -e git://github.com/Rapptz/discord.py@rewrite#egg=discord.py[voice]
discord.py
python-dotenv